+++ src/main/java/com/cse687/zirui/bookstore/order/domain/model/depreciate/DepreciationCondition.java
@@ -1,18 +1,18 @@
 package com.cse687.zirui.bookstore.order.domain.model.depreciate;
 
 public enum DepreciationCondition implements DepreciationStrategy {
-    FAIR(0.7),
-    GOOD(0.8),
-    LIKE_NEW(0.9);
+    FAIR(0.7f),
+    GOOD(0.8f),
+    LIKE_NEW(0.9f);
 
-    private final double rate;
+    private final float rate;
 
-    DepreciationCondition(double rate) {
+    DepreciationCondition(float rate) {
         this.rate = rate;
     }
 
     @Override
-    public double depreciationRate() {
+    public float depreciationRate() {
         return rate;
     }
 }
